+++ client/myapp/package.json
@@ -12,11 +12,9 @@
     "draft-js": "^0.11.7",
     "draft-js-hashtag-plugin": "^2.0.4",
     "draft-js-plugins-editor": "^3.0.0",
-    "gapi-script": "^1.2.0",
     "react": "^17.0.2",
     "react-cookie": "^4.1.1",
     "react-dom": "^17.0.2",
-    "react-google-login": "^5.2.2",
     "react-image-picker": "^1.1.1",
     "react-router-dom": "^6.2.2",
     "react-scripts": "5.0.0",
