+++ client/myapp/src/App.js
@@ -11,7 +11,10 @@ const App = () => {
     <Routes>
       <Route index element={<Main />} />
       <Route path="/mypage" element={<MyPage />} />
-      <Route path="/create/novel" element={<CreateNewNovel />} />
+      <Route
+        path="/create/novel"
+        element={<CreateNewNovel />}
+      />
       <Route path="/search" element={<SearchResult />} />
       <Route path="/novel-list/novel/:title" element={<ReaderChapter />} />
       <Route
