--- client/myapp/src/components/LoginButton.js
@@ -1,31 +1,0 @@
-import GoogleLogin from "react-google-login";
-import GlobalStyle from "../GlobalStyle";
-
-const LoginButton = () => {
-  const clientId =
-    "112172327061-95mqb878sgpt8t955rkkdug7mvgco8od.apps.googleusercontent.com";
-
-  const onSuccess = (res) => {
-    console.log("로그인 성공!", res.profileObj);
-  };
-
-  const onFailure = (res) => {
-    console.log("로그인 실패 :/", res);
-  };
-
-  return (
-    <>
-      <GlobalStyle />
-      <GoogleLogin
-        clientId={clientId}
-        buttonText={"로그인"}
-        onSuccess={onSuccess}
-        onFailure={onFailure}
-        cookiePolicy={"single_host_origin"}
-        isSignedIn={true}
-      />
-    </>
-  );
-};
-
-export default LoginButton;
