--- client/myapp/src/components/LogoutButton.js
@@ -1,22 +1,0 @@
-import { GoogleLogout } from "react-google-login";
-
-const clientId =
-  "112172327061-95mqb878sgpt8t955rkkdug7mvgco8od.apps.googleusercontent.com";
-
-const LogoutButton = () => {
-  const onSuccess = () => {
-    console.log("Log out successful!");
-  };
-
-  return (
-    <>
-      <GoogleLogout
-        clientId={clientId}
-        buttonText={"로그아웃"}
-        onLogoutSuccess={onSuccess}
-      />
-    </>
-  );
-};
-
-export default LogoutButton;
