+++ client/myapp/src/pages/Main.js
@@ -1,13 +1,12 @@
-import NavigationBar from "../components/NavigationBar";
+import NavigationBar from "./NavigationBar";
 import GlobalStyle from "../GlobalStyle";
-import { Link } from "react-router-dom";
 
 const Main = () => {
   return (
     <>
       <GlobalStyle />
       <NavigationBar />
-      <h1>메인페이지!</h1>
+      <h1>메인페이지</h1>
     </>
   );
 };
